requirements.txt
flask
line-bot-sdk
gunicorn
python-dotenv
google-generativeai
gunicorn
